backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.1
python-dotenv==1.0.1
openai==1.52.0
langchain==0.2.16
langchain-openai==0.1.23
numpy==1.26.4
tiktoken==0.7.0; python_version < "3.13"
